+++ app/models/rails_admin/action_not_allowed_decorator.rb
@@ -1,0 +1,8 @@
+module RailsAdmin
+  ActionNotAllowed.class_eval do
+
+    def initialize(msg = 'Action not allowed')
+      super
+    end
+  end
+end
